requirements.txt
opencv-python
anthropic
python-dotenv
httpx
flask
numpy
rich
discord.py
ultralytics